Directory with an empty pyproject.toml files is built into an sdist+wheel by setuptools 64 Fixes #3511
